LayoutTests/platform/ios/imported/w3c/web-platform-tests/css/css-text/i18n/ja/css-text-line-break-ja-pr-normal-expected.txt


FAIL 00B1  PLUS-MINUS SIGN may NOT appear at line start if ja and normal assert_approx_equals: expected 89.19000244140625 +/- 1 but got 59.19000244140625
FAIL 20AC  EURO SIGN may NOT appear at line start if ja and normal assert_approx_equals: expected 91.44000244140625 +/- 1 but got 61.44000244140625
FAIL 2116  NUMERO SIGN may NOT appear at line start if ja and normal assert_approx_equals: expected 99 +/- 1 but got 69
FAIL FE69  SMALL DOLLAR SIGN may NOT appear at line start if ja and normal assert_approx_equals: expected 99 +/- 1 but got 69
FAIL FF04  FULLWIDTH DOLLAR SIGN may NOT appear at line start if ja and normal assert_approx_equals: expected 99 +/- 1 but got 69
FAIL FFE1  FULLWIDTH POUND SIGN may NOT appear at line start if ja and normal assert_approx_equals: expected 99 +/- 1 but got 69
FAIL FFE5  FULLWIDTH YEN SIGN may NOT appear at line start if ja and normal assert_approx_equals: expected 99 +/- 1 but got 69
FAIL FFE6  FULLWIDTH WON SIGN may NOT appear at line start if ja and normal assert_approx_equals: expected 99 +/- 1 but got 69
00B1
文文文文文文±字字
文文文文文
文±字字
20AC
文文文文文文€字字
文文文文文
文€字字
2116
文文文文文文№字字
文文文文文
文№字字
FE69
文文文文文文﹩字字
文文文文文
文﹩字字
FF04
文文文文文文＄字字
文文文文文
文＄字字
FFE1
文文文文文文￡字字
文文文文文
文￡字字
FFE5
文文文文文文￥字字
文文文文文
文￥字字
FFE6
文文文文文文￦字字
文文文文文
文￦字字
